requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1.7
django-debug-toolbar==5.0.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-nested-routers==0.94.1
iniconfig==2.1.0
mysqlclient==2.2.7
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.5
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.1
